backend/requirements.txt
fastapi
uvicorn[standard]
supabase
google-generativeai
requests
